requirements.ras-pi.txt
--index-url https://pypi.org/simple
# Note: RPi.GPIO will not install on MacOS
RPi.GPIO